(feat) derivar el conteo de escuchas de la BD y personalizar el feed
Quita los contadores numéricos Song/Album/Artist.listeners y deriva la
popularidad (canciones, álbumes, artistas y géneros) por consulta sobre
la tabla de eventos user_song_listen. El streaming ya no incrementa
contadores: sólo registra la escucha. SongDTO.listeners pasa a ser un
conteo derivado (mapa agrupado para evitar el N+1) y AlbumDTO deja de
exponerlo.

El feed del home se personaliza por usuario: recommendArtistsForUser
toma sus artistas más escuchados (de user_song_listen) y completa con la
popularidad global sin repetir; un usuario sin historial recibe la
popularidad global (cold-start).

+++ src/main/java/anton/davila/selfpotify/controllers/dto/SongDTO.java
@@ -18,6 +18,8 @@ public class SongDTO {
     private String picture_url;
     private List<Long> artistIds;
     private List<String> artistNames;
-    private int listeners;
+    // Popularidad DERIVADA: número de escuchas de la canción contado sobre la
+    // tabla de eventos user_song_listen (no es un campo almacenado en Song).
+    private long listeners;
     private Long albumId;
 }
